feat: move age_at_vital_status to cde harmonized section
